+++ changelog.md
@@ -1,3 +1,19 @@
+# 0.0.6
+
+### Bug Fixes
+
+* Could not use 'daisy' with sample rate setter.
+
+### New Features
+
+* Add function in utilities for making daisy packets.
+* Add code to `getChannelDataArray` for ganglion and daisy data being routed over wifi
+* Create idea of protocols i.e. `BLE`, `Wifi`, and `Serial`
+
+### Breaking changes
+
+* `getChannelDataArray` now takes object as only arg.
+
 # 0.0.5
 
 ### Bug Fixes